Add advanced configuration and migration files
Several new configuration and migration files have been added to the project. This includes the creation of files related to the 'bolt-pro' package and the 'Form' module. These files contribute to adjusting various settings and establish new database tables and fields. A set of changes also include the introduction of a new 'AdvanceDate' and 'Alert' fields. This is part of a broader effort to enhance the project's functionality and customization options.